Refactor serviceWorker source into multiple files
