+++ ts/packages/agents/browser/src/extension/serviceWorker/messageHandlers.ts
@@ -1,0 +1,166 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import { getActiveTab } from "./tabManager";
+import { getTabHTMLFragments, getTabAnnotatedScreenshot } from "./capture";
+import { getRecordedActions, clearRecordedActions, saveRecordedActions } from "./storage";
+import { sendActionToAgent, ensureWebsocketConnected } from "./websocket";
+
+/**
+ * Handles messages from content scripts and popup
+ * @param message The message received
+ * @param sender The sender of the message
+ * @returns Promise resolving to the result of handling the message
+ */
+export async function handleMessage(
+    message: any,
+    sender: chrome.runtime.MessageSender
+): Promise<any> {
+    switch (message.type) {
+        case "initialize": {
+            console.log("Browser Agent Service Worker started");
+            try {
+                const connected = await ensureWebsocketConnected();
+                if (!connected) {
+                    console.log("WebSocket connection failed on initialize");
+                }
+            } catch (error) {
+                console.error("Error during initialization:", error);
+            }
+
+            return "Service worker initialize called";
+        }
+        case "refreshSchema": {
+            const schemaResult = await sendActionToAgent({
+                actionName: "detectPageActions",
+                parameters: {
+                    registerAgent: false,
+                },
+            });
+
+            return {
+                schema: schemaResult.schema,
+                actionDefinitions: schemaResult.typeDefinitions,
+            };
+        }
+        case "registerTempSchema": {
+            const schemaResult = await sendActionToAgent({
+                actionName: "registerPageDynamicAgent",
+                parameters: {
+                    agentName: message.agentName,
+                },
+            });
+
+            return { schema: schemaResult };
+        }
+        case "getIntentFromRecording": {
+            const schemaResult = await sendActionToAgent({
+                actionName: "getIntentFromRecording",
+                parameters: {
+                    recordedActionName: message.actionName,
+                    recordedActionDescription: message.actionDescription,
+                    recordedActionSteps: message.steps,
+                    existingActionNames: message.existingActionNames,
+                    fragments: message.html,
+                    screenshots: message.screenshot,
+                },
+            });
+
+            return {
+                intent: schemaResult.intent,
+                intentJson: schemaResult.intentJson,
+                actions: schemaResult.actions,
+                intentTypeDefinition: schemaResult.intentTypeDefinition,
+            };
+        }
+        case "startRecording": {
+            const targetTab = await getActiveTab();
+            if (targetTab?.id) {
+                await chrome.tabs.sendMessage(
+                    targetTab.id,
+                    {
+                        type: "startRecording",
+                    },
+                    { frameId: 0 }, // Limit action recording to the top frame for now
+                );
+            }
+            return {};
+        }
+        case "stopRecording": {
+            const targetTab = await getActiveTab();
+            if (targetTab?.id) {
+                const response = await chrome.tabs.sendMessage(
+                    targetTab.id,
+                    {
+                        type: "stopRecording",
+                    },
+                    { frameId: 0 },
+                );
+                return response;
+            }
+            return {};
+        }
+        case "takeScreenshot": {
+            const screenshotUrl = await chrome.tabs.captureVisibleTab({
+                format: "png",
+            });
+
+            return screenshotUrl;
+        }
+        case "captureHtmlFragments": {
+            const targetTab = await getActiveTab();
+            if (targetTab) {
+                const htmlFragments = await getTabHTMLFragments(targetTab);
+                return htmlFragments;
+            }
+            return [];
+        }
+        case "saveRecordedActions": {
+            await saveRecordedActions(
+                message.recordedActions,
+                message.recordedActionPageHTML,
+                message.recordedActionScreenshot,
+                message.actionIndex,
+                message.isCurrentlyRecording
+            );
+            return {};
+        }
+        case "recordingStopped": {
+            await saveRecordedActions(
+                message.recordedActions,
+                message.recordedActionPageHTML,
+                message.recordedActionScreenshot,
+                message.actionIndex,
+                false
+            );
+            return {};
+        }
+        case "getRecordedActions": {
+            const result = await getRecordedActions();
+            return result;
+        }
+        case "clearRecordedActions": {
+            try {
+                await clearRecordedActions();
+            } catch (error) {
+                console.error("Error clearing storage data:", error);
+            }
+            return {};
+        }
+        case "downloadData": {
+            const jsonString = JSON.stringify(message.data, null, 2);
+            const dataUrl =
+                "data:application/json;charset=utf-8," +
+                encodeURIComponent(jsonString);
+
+            chrome.downloads.download({
+                url: dataUrl,
+                filename: message.filename || "schema-metadata.json",
+                saveAs: true,
+            });
+            return {};
+        }
+        default:
+            return null;
+    }
+}
